Move appearance count methods to PropositionsDao
