+++ howdju-service-common/lib/daos/PropositionsDao.ts
@@ -25,10 +25,10 @@ import {
 } from "howdju-common";
 
 import { toProposition } from "./orm";
-import { normalizeText, toIdString } from "./daosUtil";
+import { normalizeText, toCountNumber, toIdString } from "./daosUtil";
 import { DatabaseSortDirection } from "./daoModels";
 import { Database, EntityNotFoundError, PropositionRow, UsersDao } from "..";
-import { keyBy, toString, uniq } from "lodash";
+import { keyBy, reduce, toString, uniq } from "lodash";
 
 export class PropositionsDao {
   constructor(
@@ -352,5 +352,44 @@ export class PropositionsDao {
       userId,
     ]);
     return result;
+  }
+
+  async readAppearanceCountForPropositionId(
+    propositionId: EntityId
+  ): Promise<number> {
+    const countsById = await this.readAppearanceCountForPropositionIds([
+      propositionId,
+    ]);
+    return countsById[propositionId];
+  }
+
+  async readAppearanceCountForPropositionIds(propositionIds: string[]) {
+    const { rows } = await this.database.query<{
+      proposition_id: number;
+      count: string;
+    }>(
+      "readAppearanceCountForPropositionIds",
+      `
+        select
+            proposition_id
+          , count(*) as count
+        from
+          propositions p join appearances a using (proposition_id)
+        where
+              p.proposition_id = any ($1)
+          and p.deleted is null
+          and a.deleted is null
+        group by proposition_id
+      `,
+      [propositionIds]
+    );
+    return reduce(
+      rows,
+      (acc, { proposition_id, count }) => {
+        acc[toIdString(proposition_id)] = toCountNumber(count);
+        return acc;
+      },
+      {} as Record<string, number>
+    );
   }
 }
